remove game files and update imports
