+++ bin/brain-calc.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-import startGame from '../src/brain-calc-game.js';
+import startGame from '../src/games/brain-calc-game.js';
 
 startGame();
