+++ bin/brain-prime.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-import startGame from '../src/brain-prime-game.js';
+import startGame from '../src/games/brain-prime-game.js';
 
 startGame();
