+++ bin/brain-progression.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-import startGame from '../src/brain-progression.js';
+import startGame from '../src/games/brain-progression.js';
 
 startGame();
