+++ src/games/brain-even-game.js
@@ -1,6 +1,6 @@
-import getRandomNumber from './helpers/get-random.js';
-import isEven from './helpers/is-even.js';
-import startGame from './game-engine.js';
+import getRandomNumber from '../helpers/get-random.js';
+import isEven from '../helpers/is-even.js';
+import startGame from '../game-engine.js';
 
 const getQuestionAndAnswer = () => {
   const question = getRandomNumber(1, 100);
